gasolina/gasolina_back/gasoline/api/tipo_combustible_viewset.py
from rest_framework import serializers, viewsets

from gasoline.models import TipoCombustible


class TipoCombustibleSerializer(serializers.ModelSerializer):
    class Meta:
        model = TipoCombustible
        fields = '__all__'


class TipoCombustibleViewSet(viewsets.ModelViewSet):
    queryset = TipoCombustible.objects.all()
    serializer_class = TipoCombustibleSerializer
